Updated mkinitramfs.sh to create devices using MAKEDEV

+++ mkinitramfs.sh
@@ -33,43 +33,21 @@ if [ -z "$1" ]; then
 	exit 1
 fi
 
+MAKEDEV="`pwd`/MAKEDEV"
+if [ ! -f "$MAKEDEV" ]; then
+	echo "Can't find MAKEDEV" 1>&2
+	exit 1
+fi
+
 install -m 0755 -d "$1"/{dev,boot,proc} || exit 1
 install -m 0755 src/kexec-loader "$1/sbin/kexec-loader"
 ln -sf "sbin/kexec-loader" "$1/init"
 
-if [ ! -e "$1/dev/console" ]; then
-	mknod -m 0600 "$1/dev/console" c 5 1 || exit 1
-fi
-if [ ! -e "$1/dev/ttyS0" ]; then
-	mknod -m 0600 "$1/dev/ttyS0" c 4 64 || exit 1
-fi
-if [ ! -e "$1/dev/ttyS1" ]; then
-	mknod -m 0600 "$1/dev/ttyS1" c 4 65 || exit 1
-fi
-if [ ! -e "$1/dev/ttyS2" ]; then
-	mknod -m 0600 "$1/dev/ttyS2" c 4 66 || exit 1
-fi
-if [ ! -e "$1/dev/ttyS3" ]; then
-	mknod -m 0600 "$1/dev/ttyS3" c 4 67 || exit 1
-fi
-if [ ! -e "$1/dev/fd0" ]; then
-	mknod -m 0600 "$1/dev/fd0" b 2 0 || exit 1
-fi
-if [ ! -e "$1/dev/fd1" ]; then
-	mknod -m 0600 "$1/dev/fd1" b 2 1 || exit 1
-fi
-if [ ! -e "$1/dev/sda" ]; then
-	mknod -m 0600 "$1/dev/sda" b 8 0 || exit 1
-fi
-if [ ! -e "$1/dev/sdb" ]; then
-	mknod -m 0600 "$1/dev/sdb" b 8 16 || exit 1
-fi
-if [ ! -e "$1/dev/sdc" ]; then
-	mknod -m 0600 "$1/dev/sdc" b 8 32 || exit 1
-fi
-if [ ! -e "$1/dev/sdd" ]; then
-	mknod -m 0600 "$1/dev/sdd" b 8 48 || exit 1
-fi
+cd "$1/dev"
+$MAKEDEV console
+$MAKEDEV ttyS{0,1,2,3}
+$MAKEDEV sd{a,b,c,d,e,f,g,h}
+$MAKEDEV hd{a,b,c,d,e,f,g,h}
 
 echo "Created initramfs in $1"
 echo "Be sure to set the full path of $1 in CONFIG_INITRAMFS_SOURCE"
